Cleanup: IOS OSPF db rtr changed state names
Changed state names to better align with RFC-2328:
 - LSAs is now Link
 - LSAInfo is now LSA_Header

The name LSAs for the state in the previous commit was misleading. Per
RFC-2328, the LSA includes the link data, which was captured by the
state previously known as LSAs, as well as the LSA header, which was
captured by the state previously known as LSAInfo.