Cleanup: IOS OSPF router LSA links are now lists
cisco_ios: show ip ospf database router
Link attributes are now lists within the context of the parent LSA.
Each row of data is now an LSA that contains link data.  This is more in
line with the nature of the output of the IOS command, and also more
acccurately depicts the LSA data structure as defined in RFC-2328.