Enhancement: IOS capture Bit E in router LSA
cisco_ios: show ip ospf database router

If a router withiin the OSPF topology is configured to redistribute into
OSPF or to originate a default route into OSPF, it is an Autonomous
System Boundary Router (ASBR). Per RFC-2328, a router must indicate
whether it is an ASBR by setting "Bit E" (for External) in its router
LSA.

If Bit E is enabled in a router LSA, the output of the `show ip ospf
database router` command on Cisco IOS will include an extra line within
the output of the LSA header information that reads, "AS Boundary
Router".  This commit accounts for the possible existence of this line
and modifies the data structure by `Value LSA_ASBR`.  If the router is
an ASRB, `LSA_ASBR` will have a value of `AS Boundary Router`, otherwise
it will be an empty string.